isntagram/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.8
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
colorama==0.4.4
debugpy==1.6.2
decorator==5.1.0
defusedxml==0.7.1
Django==3.2.13
django-appconf==1.0.5
django-bootstrap5==22.1
django-extensions==3.1.3
django-imagekit==4.1.0
djangorestframework==3.14.0
entrypoints==0.4
executing==1.0.0
Faker==13.15.1
fastjsonschema==2.16.1
idna==3.3
ipykernel==6.15.1
ipython==7.27.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.1.2
jsonschema==4.7.2
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
packaging==21.3
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
pilkit==2.0
Pillow==9.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.20
psutil==5.9.1
pure-eval==0.2.2
pycparser==2.21
Pygments==2.10.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.4
pywin32==304
pywinpty==2.0.6
pyzmq==23.2.0
requests==2.28.1
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.5.0
terminado==0.15.0
tinycss2==1.1.1
tornado==6.2
traitlets==5.1.0
tzdata==2022.2
urllib3==1.26.11
wcwidth==0.2.5
webencodings==0.5.1
